Auto offline enhancement
Added logic to use the new poe.xyz offline feature that expands on the
previous change that Kapps made. The user will only be set as online if
POE is running and they haven't been AFK for more than 10 minutes. The
enhancement is that once POE is closed or the user goes AFK then they
are set to be offline. The main online timer was set to run at 1 minute
to help with this, and extra logic was added so that we don't send
multiple repeated online or offline commands to poe.xyz.

File needs to be cleaned up before merge since I reformatted it to my
normal style.