Fixed re-online bug
Fixed the bug that would prevent the user from staying online for more
than the first hour. The last time that we registered as online with
poe.trade is now retained, and after 55 minutes (5 minute backoff so
that the 1 minute timer has time to catch and we don't have any
offline/online scenarios) we reregister for another hour.