More completely find items for same name recipe.
The previous version would, among all items with the same name, only
create one recipe result.  For the two-item version, this would mean the
recipe would be shown only if there were exactly two items with that
same name.  For the three-item version, this would mean the resulting
recipe would include every item with that name.  Both of these cases are
undesirable.

The new version will continue creating recipe results from the set of
items with the same name until all of the items are included in a recipe
result, where that recipe result might have the exact right number of
items for the recipe, or might be short some items.