Add DescriptiveName property to Item class.
This is meant to provide a short but informative string description of
the time.  By default, it provides the quality (if present) and item
level of the item, along with the type name.  Gear and AbyssJewel
subclasses provide additional information, as they have a rarity, along
with a special name when they have a rarity other than Normal.