Fix all misspellings of "Abyss Jewel".
According to the official Abyss League announcement, the correct term
for the new jewel type is "Abyss Jewel".  However, the new item socket
type these can be inserted into is an "Abyssal Socket", items with these
sockets are "Abyss-socketed", new uniques are "Abyss Uniques", and when
an Abyss expands it could lead to the "Abyssal Depths".