Sort compatible types to use longer base types.
There are base types in the same gear type where one base type is a
non-trivial substring of another base type.  For example, there are
"Tricone", "Noble Tricorne", and "Sinner Tricorne" base types.  Since we
determine and item's base type by checking if each base type is a
substring of that item's type line (since the type line can include
affixes), we need to make sure we do not use a shorter base type when a
longer one is more correct.  The simple fix used here, after building
the list of compatible types for a gear type, sort that list by string
length, so the longer strings are checked first.