Add support for full bestiary orbs.
A full bestiary orb contains a best that can be added to your
bestiary.  The adjective "full" is from the file name of the item
icon (whereas bestiary orbs without a beast are "empty").  Full bestiary
orbs have unique properties Genus, Group, and Family, which describe the
type of beast contained inside.  They are also another type of item that
has Rarity.