Handle getting a property from a null list.
Pantheon Soul items, despite being a currency item, do not have any
properties, so when we try to determine the stack info from the item's
properties, we want to gracefully fall back to a default.

This also adds "PantheonSoul" as an OrbType, and collapses all the
various captured souls into that single OrbType.