Add missing item and rearrange item lists.
The major changes include adding in new maps and divination cards.  Most
of the maraketh weapons and map-only base types were also previously
missing.

For completeness' sake, each release of the end-game maps is listed in
the maps gear section, annotated with release and tier levels.  This
leads to some duplication, so only distinct items are used when reading
the data file.

Also, the PoE wiki claims there are currently 201 divination cards, but
only 200 are provided in the list of cards.  This might be a technical
issue, since the page's category indicates the table has reached a "hard
limit", but so far I can not deduce which card is missing, if any.