Add an option to disable refreshing data on login.
If this value is true, the login process behaves as it did before.  If
it is false, only stash tabs that are missing are downloaded.  Thus, a
new user can set this value to false and all of his tabs will be
downloaded during the first login.  Moreover, if only some tabs have
been downloaded (e.g., the user closes Procurement while downloading
tabs), only the tabs that were not downloaded yet will be downloaded.