Fake mapping divination cards onto a quad tab.
The location for divination cards on a divination card tab is a line --
the Y value is always zero, and there is a unique X value used for each
type of divination card.  Map this linear series of locations onto the
planar quad stash tab, so there is a location to display all divination
cards.  The normal stash tab does not have enough locations to display
all divination cards.