Use distict conditions for SBT recipe variants.
Not considering Shaper/Elder status, the most specific type of
SameBaseTypeRecipe is that all of the items have 20% quality and are
unidentified.  Other variants require only one of those conditions, or
neither to be true.  In game, a recipe can be completed if some of the
items have "extra" qualitys.  For example, if two items are identified
but the third is unidentified, then they can be used for the
unidentified variant of the SBT recipe.

However, it is subjectively preferable to use exactly the worst items
possible for each recipe.  Thus, for the variant that can take
identified or unidentified items, use *only* identified (or normal)
items.  Then, unidentified items are "saved" for the better variant.

This change, therefore, creates additional constraints that look for
these artificially stricter conditions, and presents only those are
potential results.