Check for a null ShopSettings value before using.
It might not be a likely situation, but we could be attempting to
dereference null objects while reconstructing the ShopSettings settings
element.  If we find one of the settings in null, consider the entire
save action a failure, and notify the user.  Non-null settings are still
saved, so this is overkill.  It would be better to abort the save and
not change the Settings.xml file at all, but that is not simple to do at
the moment.