Use empty strings for null shop thread info.
If the user has not modified any of the shop thread information for a
league, it looks like the threadId and threadTitle will be null.  Trying
to set these as the attribute values for the ShopSettings' leagues'
attributes triggers an exception, so when these values are null, use
empty strings instead.