Update the model and transports for CloudFlare.
In order to authenticate using Session ID when the PoE website is behind
CloudFlare, requests need to provide additional cookies, and use the
same user agent as the browser which obtained those cookies.  There are
also some additional headers that need to be provided (though perhaps
not *all* of them are essential).  It also looks like some responses are
gzip-encoded (at least the request for the account name is), so we
should enable automatic decompression.  The HttpTransport uses the same
user agent as it did before by default, but that can be overridden now,
and that user agent is used for all requests after it is set.