Fix crafting slot positioning in currency tab
This commit fixes the placement of items in the crafting slot by setting
the scale on the item style. The caveat with this is that because the
scaling is on the xaml, it does not apply to the sockets and links,
thus the link images overlap with some of the sockets.