Add a setting to use short recipe descriptions.
If this setting, "UseShortRecipeDisplayDescriptions", is false, the
recipe descriptions are displayed as before.  If it is true,
shorter (but more cryptic) descriptions are used, which allow the number
of matches to be visible.  This currently only applies to the
SameBaseTypeRecipe descriptions, but could be added to other long recipe
descriptions.