More gracefully handle exceeding the API limit.
If another tool that uses the GGG API is used alongside Procurement, it
is very easy to exceed the API limit.  This change more gracefully
handles that case, but is very cautious about it.  If an overflow is
detected, completely fill up the window of active tasks, and wait for at
least as long as the window size.