Update URLs to use for character tab images.
The character tab selectors use hard-coded URLs for their composite
images, since they are fake tabs that do not get actual URLs via the GGG
API.  It seems the URLs for these images (black-colored tabs) changed
with the last league, so they need to be updated.

A better solution would be to dynamically determine the URLs based on
the URLs of other tabs' images, but that is a larger and more involved
change that can be done later.