Add a class to watch and read the client log file.
Right now, this ClientLogFileWatcher both listens for file changes and
reads the file on a tight schedule, in order to minimize the amount of
time between the file being updated and Procurement seeing it.  Using a
FileSystemWatcher is, unfortunately, not sufficient, since this
generally misses the updates PoE makes to the file.

Currently the only log message that is searched for is for when the user
changes areas.  The file watcher then triggers an event, which can be
listened for by other parts of Procurement.